# docs/openapi.yaml
openapi: 3.0.3
info:
  title: APEX Router API
  description: |
    APEX is a pre-alpha autonomous agent platform. 
    
    ## Authentication
    All API requests require HMAC signature authentication:
    - `X-APEX-Signature`: HMAC-SHA256 signature
    - `X-APEX-Timestamp`: Unix timestamp (within 5 minutes)
    
    Set `APEX_AUTH_DISABLED=1` for local development.
    
    ## Base URL
    `http://localhost:3000`
  version: 1.3.0
  contact:
    name: APEX Team
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: http://localhost:3001
    description: Gateway (proxied)
...
paths:
...
  /api/v1/skills:
    get:
      summary: List skills
      description: List all registered skills
      tags: [Skills]
      responses:
        '200':
          description: List of skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Skill'
    post:
      summary: Register skill
      description: Register a new skill
      tags: [Skills]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - version
                - tier
              properties:
                name:
                  type: string
                version:
                  type: string
                tier:
                  type: string
                  enum: [T0, T1, T2, T3]
  /api/v1/skills/{name}:
    get:
      summary: Get skill
      description: Get skill details by name
      tags: [Skills]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Skill details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
    delete:
      summary: Delete skill
      description: Delete a skill by name
      tags: [Skills]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Skill deleted
...
  /api/v1/skills/execute:
    post:
      summary: Execute skill
      description: Execute a skill directly
      tags: [Skills]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - skill_name
                - input
              properties:
                skill_name:
                  type: string
                input:
                  type: object
...
  /api/v1/memory/reflections:
    get:
      summary: Get reflections
      tags: [Memory]
      responses:
        '200':
          description: List of reflections
    post:
      summary: Add reflection
      tags: [Memory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
...
components:
  schemas:
...
    Skill:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        tier:
          type: string
        description:
          type: string
        healthy:
          type: boolean
        last_health_check:
          type: string
...